v1.9 complete — deterministic forward return engine & foundation closure

+++ src/main/java/com/trading/scanner/repository/ScanResultRepository.java
@@ -7,7 +7,7 @@ import java.time.LocalDate;
 import java.util.List;
 
 @Repository
-public interface ScanResultRepository extends JpaRepository<ScanResult, Long> {
+public interface ScanResultRepository extends JpaRepository<ScanResult, Integer> {
     
     List<ScanResult> findTop10ByOrderByScanDateDesc();
